Add documentation comments in the main additional_fonts.txt Move a large part of the font handling code (finding and registering fonts) to a separate python file, the main file is far too big
